+++ src/components/LinkModals.tsx
@@ -11,7 +11,7 @@ import getSocialNetworks from "../utils/getSocialNetworks"
 
 const LinkModalBG: React.FC<React.PropsWithChildren> = ({ children }) => {
     return (
-        <div id="LinkButtonSelectedBG" className="fixed z-10 w-full h-screen backdrop-blur-[2px] duration-200" style={{ display: 'none', backgroundColor: 'rgba(0,0,0,0)', opacity: '0%' }}>
+        <div id="LinkButtonSelectedBG" className="fixed z-10 w-full h-screen backdrop-blur-[2px] duration-200 items-center" style={{ display: 'none', backgroundColor: 'rgba(0,0,0,0)', opacity: '0%' }}>
             {children}
         </div>
     )
@@ -51,8 +51,42 @@ const LinkModalOptions: React.FC<{ type: string, lang: string }> = ({ type, lang
     const OpenTab = () => {
         window.open(getSocialNetworks(type, 'link'), "_blank")
     }
-    const CopyToClipboard = () => {
+    const CopyToClipboard = async (copyText: 'default' | 'email' = 'default') => {
+        const copyUserText = document.getElementById("copyUserText") as HTMLSpanElement
         navigator.clipboard.writeText(getSocialNetworks(type, 'user'))
+            .then(() => {
+                copyUserText.classList.add('animate-copytext')
+                setTimeout(() => {
+                    copyUserText.style.fontSize = "0.8rem"
+                    copyUserText.style.color = 'rgba(238, 238, 238, 0.8)'
+                    if (copyText === 'default') {
+                        copyUserText.innerHTML = getLocaleString('copy_successfully', lang, true)
+                    } else if (copyText === 'email') {
+                        copyUserText.innerHTML = getLocaleString('copy_email_successfully', lang, true)
+                    }
+                }, 100)
+                setTimeout(() => {
+                    copyUserText.classList.remove('animate-copytext')
+                }, 200)
+                setTimeout(() => {
+                    copyUserText.classList.add('animate-copytext')
+                }, 800)
+                setTimeout(() => {
+                    copyUserText.style.fontSize = "1.125rem"
+                    copyUserText.style.color = 'rgba(238, 238, 238)'
+                    if (copyText === 'default') {
+                        copyUserText.innerHTML = getLocaleString(`copy_user`, lang)
+                    } else if (copyText === 'email') {
+                        copyUserText.innerHTML = getLocaleString(`copy_email`, lang)
+                    }
+                }, 900)
+                setTimeout(() => {
+                    copyUserText.classList.remove('animate-copytext')
+                }, 1000)
+            })
+            .catch((reason) => {
+                console.log(reason)
+            })
     }
 
     if (type) {
@@ -60,22 +94,31 @@ const LinkModalOptions: React.FC<{ type: string, lang: string }> = ({ type, lang
             <>
                 {
                     type === 'facebook' || type === 'twitter' || type === 'instagram' || type === 'twitch' || type == 'linkedin' || type === 'github' || type === 'soundcloud' ? (
-                        <button className="font-light text-lg w-full py-3 bg-transparent hover:bg-[#ffffff22] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`access_link`, lang)}</button>
+                        <button className="font-light text-lg w-full h-12 bg-transparent hover:bg-[#ffffff0f] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`access_link`, lang)}</button>
                     ) : null
                 }
                 {
                     type === 'ngl' ? (
-                        <button className="font-light text-lg w-full py-3 bg-transparent hover:bg-[#ffffff22] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`ngl_link`, lang)}</button>
+                        <button className="font-light text-lg w-full h-12 bg-transparent hover:bg-[#ffffff0f] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`ngl_link`, lang)}</button>
                     ) : null
                 }
                 {
                     type === 'gmail' ? (
-                        <button className="font-light text-lg w-full py-3 bg-transparent hover:bg-[#ffffff22] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`send_email`, lang)}</button>
+                        <button className="font-light text-lg w-full h-12 bg-transparent hover:bg-[#ffffff0f] duration-100" onClick={ev => OpenTab()} >{getLocaleString(`send_email`, lang)}</button>
                     ) : null
                 }
                 {
                     type === 'twitter' || type === 'instagram' || type === 'twitch' || type === 'discord' || type === 'github' || type === 'soundcloud' ? (
-                        <button className="font-light text-lg w-full py-3 bg-transparent hover:bg-[#ffffff22] duration-100" onClick={ev => CopyToClipboard()} >{getLocaleString(`copy_user`, lang)}</button>
+                        <button className="flex font-light text-lg w-full h-12 items-center bg-transparent hover:bg-[#ffffff0f] duration-100 justify-center" onClick={ev => CopyToClipboard()} >
+                            <span id='copyUserText' className="flex">{getLocaleString(`copy_user`, lang)}</span>
+                        </button>
+                    ) : null
+                }
+                {
+                    type === 'gmail' ? (
+                        <button className="flex font-light text-lg w-full h-12 items-center bg-transparent hover:bg-[#ffffff0f] duration-100 justify-center px-6" onClick={ev => CopyToClipboard('email')} >
+                            <span id='copyUserText' className="flex">{getLocaleString(`copy_email`, lang)}</span>
+                        </button>
                     ) : null
                 }
             </>
@@ -88,7 +131,7 @@ const LinkModalOptions: React.FC<{ type: string, lang: string }> = ({ type, lang
 const LinkModalClose: React.FC<{ action: React.Dispatch<React.SetStateAction<boolean>>, lang: string }> = ({ action, lang }) => {
 
     return (
-        <button className="text-red-500 text-lg w-full tablet:w-[24rem] bg-[#00000088] py-3 rounded-md border border-red-500 hover:bg-[#ffffff22] duration-100" onClick={ev => action(false)} >{getLocaleString('close', lang)}</button>
+        <button className="text-red-500 text-lg w-full tablet:w-[24rem] bg-[#00000088] h-12 rounded-md border border-red-500 hover:bg-[#000000dd] duration-100" onClick={ev => action(false)} >{getLocaleString('close', lang)}</button>
     )
 }
 
@@ -123,8 +166,8 @@ const LinkModals = () => {
 
     return (
         <LinkModalBG>
-            <div id="LinkButtonSelected" className="fixed items-center mt-auto font-normal pb-6 px-4 w-full h-screen duration-200" style={{ top: '100%' }}>
-                <div className="w-full tablet:w-[24rem] mb-2 border bg-[#00000088] rounded-md pt-3 border-main-white/50 mt-auto items-center">
+            <div id="LinkButtonSelected" className="fixed font-normal h-screen px-4 w-full duration-200 justify-center items-center" style={{ top: '100%' }}>
+                <div className="w-full tablet:w-[24rem] mb-2 border bg-[#00000088] rounded-md pt-3 border-main-white/50 items-center">
                     <LinkModalIcons type={Context.linkSelection} />
                     <LinkModalDesc type={Context.linkSelection} lang={Context.lang} />
                     <div className="bg-[#333] w-full h-[1px]" />