+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## December 21st, 2022: *(0.2.0)*
+- Link modals were vertically centered on-screen to avoid mobile buttons to appear above the "Close" button.
+- "Copy to Clipboard" now has a reaction animation when the "Copy User to Clipboard" button is clicked.
+- Added "Copy E-mail Address" option for Gmail link.
+
 ## December 20th, 2022: *(0.1.1)*
 - Better components structure.
 - "Copy User to Clipboard" functionality now works correctly.